django_projects/device_tracker_services/device_tracker_backend/collector/serializers.py
from rest_framework.serializers import (
    ModelSerializer,
)

from collector.models import Network


class NetworkSerializer(ModelSerializer):

    class Meta:
        model = Network
        field = []
