django_projects/device_tracker_services/device_tracker_backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.6
